NEWRELIC-5282 consolidated the promise and when.js instrumentation as it is the only place it is being used. I also extracted out code into their own functions to reduce complexity. lastly I moved out some standalone classes around Contextualizer and the spec constant to make it easier to navigate
