+++ lib/instrumentation/when/constants.js
@@ -1,11 +1,9 @@
 /*
- * Copyright 2020 New Relic Corporation. All rights reserved.
+ * Copyright 2022 New Relic Corporation. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
 'use strict'
-
-const promInit = require('./promise')
 
 const STATIC_PROMISE_METHODS = [
   'reject',
@@ -21,12 +19,22 @@ const STATIC_PROMISE_METHODS = [
 
 const WHEN_SPEC = {
   name: 'when',
+  // name of the property where the Promise constructor lives
   constructor: 'Promise',
+  // wrap The Promise constructor
   executor: true,
+  /**
+   * The mapping for Promise instance method concepts (i.e. `then`). These are
+   * mapped on the Promise class' prototype.
+   */
   $proto: {
     then: ['then', 'done', 'spread', 'finally', 'ensure'],
     catch: ['catch', 'otherwise']
   },
+  /**
+   * The mapping for Promise static method concepts (i.e. `settle`, `race`). These
+   * are mapped on the Promise class itself.
+   */
   $static: {
     cast: STATIC_PROMISE_METHODS,
     $copy: STATIC_PROMISE_METHODS.concat([
@@ -49,15 +57,10 @@ const WHEN_SPEC = {
       'onPotentiallyUnhandledRejection'
     ]
   },
+  // The mapping for library-level static method concepts (i.e. `reject`, `resolve`).
   $library: {
     cast: STATIC_PROMISE_METHODS
   }
 }
 
-module.exports = function initialize(agent, library) {
-  if (!library || !library.Promise) {
-    return false
-  }
-
-  promInit(agent, library, WHEN_SPEC)
-}
+module.exports.WHEN_SPEC = WHEN_SPEC
