+++ lib/instrumentations.js
@@ -37,7 +37,7 @@ module.exports = function instrumentations() {
     'superagent': { module: '@newrelic/superagent' },
     'undici': { type: MODULE_TYPE.TRANSACTION },
     '@hapi/vision': { type: MODULE_TYPE.WEB_FRAMEWORK },
-    'when': { type: null },
+    'when': { module: './instrumentation/when' },
     'winston': { type: MODULE_TYPE.GENERIC }
   }
 }
